Fix model grid position for FCOs with position 0x7FFFFFFF (or thereabouts). Works, but technically lots of UB with unsigned overflow elsewhere
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@2600 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/Gme/ModelGrid.cpp
@@ -119,7 +119,7 @@ bool CModelGrid::IsAvailableG(CPoint &pt,CSize &size)
 bool CModelGrid::IsAvailable(const CRect& rect)
 {
 	FindStartEnd(rect,0);
-	if(startx < 0 || starty < 0 || endx >= GME_MAX_GRID_DIM || endy >= GME_MAX_GRID_DIM)
+	if(startx < 0 || starty < 0 || (unsigned int)endx >= GME_MAX_GRID_DIM || (unsigned int)endy >= GME_MAX_GRID_DIM)
 		return false;
 	for(int y = starty; y <= endy; y++)
 		for(int x = startx; x <= endx; x++)
